+++ docs/ChangeLog.md
@@ -9,6 +9,7 @@
 ### Added
 
 * Monkey Patching on `exit()`. *CI PHPUnit Test* could convert `exit()` in your code to Exception on the fly. See [#32](https://github.com/kenjis/ci-phpunit-test/pull/32).
+* `$this->request->setCallablePreConstructor()` to inject mocks into your controller constructors. See [#29](https://github.com/kenjis/ci-phpunit-test/pull/29).
 
 ## v0.4.0 (2015/07/21)